requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
chromadb
tiktoken
faiss-cpu
sentence-transformers
